AutoProof.AST

Description

Defines an abstract syntax tree class and related functions.

Abstract syntax tree class and metadata type

class AST t where

Abstract syntax tree class.

Minimal complete definition

root, children, (metadata | height, size)

Associated Types

type Root t

The type of the values annotating AST nodes.

Methods

root :: t -> Root t

The value at the AST's root node.

children :: t -> [t]

The AST's child ASTs.

height :: t -> Int

Number of edges on the longest path from the root of the AST to a leaf.

size :: t -> Int

Number of nodes in the AST.

metadata :: t -> ASTMetadata

The AST's metadata

data ASTMetadata

Container type for AST properties, intended for constant-time access.

Constructors

ASTMetadata

Fields

getHeight :: Int

Get an AST's height (see also height)

getSize :: Int

Get an AST's size (see also size)

AST functions

subtrees :: (AST t, Ord t) => t -> Set t

(subtrees t) is the set of all subtrees of an AST t (including t itself).

properSubtrees :: (AST t, Ord t) => t -> Set t

(properSubtrees t) is the the set of all proper subtrees of an AST t (i.e., not including t itself).

Helper functions for creating AST constructors

atomicASTConstructor :: (ASTMetadata -> a -> t) -> a -> t

Helper function for creating metadata-aware ASTs.

unaryASTConstructor :: AST t => (ASTMetadata -> t -> t) -> t -> t

Helper function for creating metadata-aware ASTs.

binaryASTConstructor :: AST t => (ASTMetadata -> t -> t -> t) -> t -> t -> t

Helper function for creating metadata-aware ASTs.

unaryRootedASTConstructor :: AST t => (ASTMetadata -> a -> t -> t) -> a -> t -> t

Helper function for creating metadata-aware rooted ASTs.

binaryRootedASTConstructor :: AST t => (ASTMetadata -> a -> t -> t -> t) -> a -> t -> t -> t

Helper function for creating metadata-aware rooted ASTs.

ternaryRootedASTConstructor :: AST t => (ASTMetadata -> a -> t -> t -> t -> t) -> a -> t -> t -> t -> t

Helper function for creating metadata-aware rooted ASTs.