AutoProof.Classical.Proof.Glivenko

Description

Glivenko translation.

Documentation

glivenkoTranslate :: Ord a => Judgement a -> Judgement a

Translate one judgement into another such that the first is classically provable if and only if the second is intuitionistically provable. Specifically, the Glivenko translation of a judgement \(g \vdash a\) is the judgement \(g, \lnot a \vdash \bot\).

isProvableGlivenko :: Ord a => Judgement a -> Bool

Determine whether a judgement is classically valid using Glivenko's translation.