AutoProof.Classical.SAT.Baseline

Description

Naive baseline SAT algorithm.

Documentation

naiveSAT :: Eq a => Formula a -> Bool

Naive satisfiability algorithm.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[x := \top]\) is satisfiable or \(a[x := \bot]\) is satisfiable.

Examples

Expand

>>> naiveSAT $ Var "a"
True

>>> naiveSAT $ And (Var "a") (Not (Var "a"))
False

naiveSATAssignment :: Ord a => Formula a -> Maybe (Map a Bool)

Naive satisfiability algorithm, returning a satisfying truth assignment.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[x := \top]\) is satisfiable or \(a[x := \bot]\) is satisfiable.

Examples

Expand

>>> naiveSATAssignment $ And (Not (Var "a")) (Or (Var "b") (Var "c"))
Just (fromList [("a",False),("b",True),("c",True)])

>>> naiveSATAssignment $ And (Var "a") (Not (Var "a"))
Nothing