AutoProof.Classical.SAT.TruthAssignment

Description

Defines the TruthAssignment class.

Documentation

class TruthAssignment t a where

Class representing truth assignments for evalutating propositional logic formulas.

Minimal complete definition

evalVar

Methods

evalVar :: t -> a -> Bool

Evaluate the truth value of a variable under a given truth assignment.

evalFormula :: t -> Formula a -> Bool

Evaluate the truth value of a formula under a given truth assignment.

(|=) :: t -> Formula a -> Bool

Infix alias of evalFormula.

Examples

Expand

>>> t = Map.fromList $ [("a", True), ("b", False)]
>>> t |= And (Var "a") (Not (Var "b"))
True