AutoProof.Formula.Types

Description

Defines the Formula type, its constructors, and its instances.

Formula type

data Formula a where

Formulas of propositional logic are built inductively from atomic propositions

truth \(\top\),

falsity \(\bot\), and

propositional variables \(a, b, c, \ldots\)

using the unary connective

negation \(\lnot\)

and the binary connectives

implication \(\rightarrow\),

disjunction \(\lor\),

conjunction \(\land\), and

equivalence \(\leftrightarrow\).

Bundled Patterns

pattern Lit :: Bool -> Formula a
Propositional literal/constant. (Lit True) is \(\top\) (i.e., truth, tautology, or top), and (Lit False) is \(\bot\) (i.e., falsity, contradiction, or bottom).

pattern Var :: a -> Formula a
Propositional variable. (Var x) represents a propositional variable \(x\).

pattern Not :: Formula a -> Formula a
Negation. (Not p) represents the formula \(\lnot p\).

pattern Imp :: Formula a -> Formula a -> Formula a
Implication. (Imp p q) represents the formula \(p \rightarrow q\).

pattern Or :: Formula a -> Formula a -> Formula a
Disjunction. (Or p q) represents the formula \(p \lor q\).

pattern And :: Formula a -> Formula a -> Formula a
Conjunction. (And p q) represents the formula \(p \land q\).

pattern Iff :: Formula a -> Formula a -> Formula a
Equivalence. (Iff p q) represents the formula \(p \leftrightarrow q\).

Pretty-printing

prettyFormula :: PrettyPrintable a => Formula a -> String

Get a pretty-printed representation of a propositional formula.

Examples

Expand

>>> prettyFormula $ Imp (Var 'a') (Or (Var 'b') (Lit False))
"a → (b ∨ ⊥)"