AutoProof.Internal.Classical.SAT

Description

Satisfiability algorithms.

Truth assignments

class TruthAssignment t a where

Class representing truth assignments for evalutating propositional logic formulas.

Minimal complete definition

evalVar

Methods

evalVar :: t -> a -> Bool

Evaluate the truth value of a variable under a given truth assignment.

evalFormula :: t -> Formula a -> Bool

Evaluate the truth value of a formula under a given truth assignment.

(|=) :: TruthAssignment t a => t -> Formula a -> Bool infix 5

Semantic entailment relation. This is an infix alias of evalFormula.

Examples

Expand

>>> t = Map.fromList $ [("a", True), ("b", False)]
>>> t |= And (Var "a") (Not (Var "b"))
True

Satisfiability algorithms

simpleSAT :: Eq a => Formula a -> Bool

A simple baseline satisfiability algorithm.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[\top/x]\) is satisfiable or \(a[\bot/x]\) is satisfiable.

Examples

Expand

>>> simpleSAT $ Var "a"
True

>>> simpleSAT $ And (Var "a") (Not (Var "a"))
False

simpleSATAssignment :: Ord a => Formula a -> Maybe (Map a Bool)

A simple baseline satisfiability algorithm, returning a satisfying truth assignment if there is one.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[\top/x]\) is satisfiable or \(a[\bot/x]\) is satisfiable.

Examples

Expand

>>> a = And (Not (Var "a")) (Or (Var "b") (Var "c")) -- satisfiable
>>> Just t = simpleSATAssignment a
>>> t
fromList [("a",False),("b",True),("c",True)]
>>> t |= a
True

>>> simpleSATAssignment $ And (Var "a") (Not (Var "a")) -- unsatisfiable
Nothing