AutoProof.Internal.Classical.SAT.Algorithms

Description

Satisfiability algorithms.

Simple baseline algorithm

simpleSAT :: Eq a => Formula a -> Bool

A simple baseline satisfiability algorithm.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[\top/x]\) is satisfiable or \(a[\bot/x]\) is satisfiable.

Examples

Expand

>>> simpleSAT $ Var "a"
True

>>> simpleSAT $ And (Var "a") (Not (Var "a"))
False

simpleSATAssignment :: Ord a => Formula a -> Maybe (Map a Bool)

A simple baseline satisfiability algorithm, returning a satisfying truth assignment if there is one.

This algorithm is based on the observation that if \(a\) is a propositional formula and \(x\) is a propositional variable, then \(a\) is satisfiable if and only if either \(a[\top/x]\) is satisfiable or \(a[\bot/x]\) is satisfiable.

Examples

Expand

>>> a = And (Not (Var "a")) (Or (Var "b") (Var "c")) -- satisfiable
>>> Just t = simpleSATAssignment a
>>> t
fromList [("a",False),("b",True),("c",True)]
>>> t |= a
True

>>> simpleSATAssignment $ And (Var "a") (Not (Var "a")) -- unsatisfiable
Nothing