AutoProof.Internal.Formula

Description

Defines the Formula type and related functions.

Formula type

data Formula a where

Formulas of propositional logic are built inductively from atomic propositions

truth \(\top\),

falsity \(\bot\), and

propositional variables \(a, b, c, \ldots\)

using the unary connective

negation \(\lnot\)

and the binary connectives

implication \(\rightarrow\),

disjunction \(\lor\),

conjunction \(\land\), and

equivalence \(\leftrightarrow\).

Bundled Patterns

pattern Lit :: Bool -> Formula a
Propositional literal/constant. (Lit True) is \(\top\) (i.e., truth, tautology, or top), and (Lit False) is \(\bot\) (i.e., falsity, contradiction, or bottom).

pattern Var :: a -> Formula a
Propositional variable. (Var x) represents a propositional variable \(x\).

pattern Not :: Formula a -> Formula a
Negation. (Not p) represents the formula \(\lnot p\).

pattern Imp :: Formula a -> Formula a -> Formula a
Implication. (Imp p q) represents the formula \(p \rightarrow q\).

pattern Or :: Formula a -> Formula a -> Formula a
Disjunction. (Or p q) represents the formula \(p \lor q\).

pattern And :: Formula a -> Formula a -> Formula a
Conjunction. (And p q) represents the formula \(p \land q\).

pattern Iff :: Formula a -> Formula a -> Formula a
Equivalence. (Iff p q) represents the formula \(p \leftrightarrow q\).

Pretty-printing

prettyFormula :: PrettyPrintable a => Formula a -> String

Get a pretty-printed representation of a propositional formula.

Examples

Expand

>>> prettyFormula $ Imp (Var 'a') (Or (Var 'b') (Lit False))
"a → (b ∨ ⊥)"

Operations on formulas

subformulas :: Ord a => Formula a -> Set (Formula a)

Get the set of subformulas of a propositional formula.

Examples

Expand

>>> subformulas $ Or (Var 'x') (And (Var 'y') (Var 'z'))
fromList [Var 'x',Var 'y',Var 'z',And (Var 'y') (Var 'z'),Or (Var 'x') (And (Var 'y') (Var 'z'))]

substitute :: Eq a => Formula a -> a -> Formula a -> Formula a

(substitute a x p) represents \(a[x := p]\), the substitution of each occurence of the variable \(x\) in the formula \(a\) by the formula \(p\).

Examples

Expand

>>> substitute (Imp (Var 'e') (Var 'e')) 'e' (And (Var 'a') (Var 'a'))
Imp (And (Var 'a') (Var 'a')) (And (Var 'a') (Var 'a'))

getAnyVariable :: Formula a -> Maybe a

Obtain a propositional variable from a formula, if there is one.

Examples

Expand

>>> getAnyVariable (Or (Var "a") (Var "b"))
Just "a"

>>> getAnyVariable (Lit False :: Formula String)
Nothing