AutoProof.Internal.Formula.Operations

Description

Defines miscellaneous operations on formulas.

Documentation

subformulas :: Ord a => Formula a -> Set (Formula a)

Get the set of subformulas of a propositional formula.

Examples

Expand

>>> subformulas $ Or (Var 'x') (And (Var 'y') (Var 'z'))
fromList [Var 'x',Var 'y',Var 'z',And (Var 'y') (Var 'z'),Or (Var 'x') (And (Var 'y') (Var 'z'))]

substitute :: Eq a => Formula a -> a -> Formula a -> Formula a

(substitute a x p) represents \(a[x := p]\), the substitution of each occurence of the variable \(x\) in the formula \(a\) by the formula \(p\).

Examples

Expand

>>> substitute (Imp (Var 'e') (Var 'e')) 'e' (And (Var 'a') (Var 'a'))
Imp (And (Var 'a') (Var 'a')) (And (Var 'a') (Var 'a'))

getAnyVariable :: Formula a -> Maybe a

Obtain a propositional variable from a formula, if there is one.

Examples

Expand

>>> getAnyVariable (Or (Var "a") (Var "b"))
Just "a"

>>> getAnyVariable (Lit False :: Formula String)
Nothing