AutoProof.Internal.Judgement

Description

Defines the Judgement type and related functions.

Types and constructors

type Context a = Set (Formula a)

A set of propositional formulas, used as antecedents of a judgement.

data Judgement a

(Judgement c a) represents the judgement or sequent \(c \vdash a\).

>>> Judgement Set.empty (Imp (And (Var 'a') (Var 'b')) (Var 'a'))
[] |- Imp (And (Var 'a') (Var 'b')) (Var 'a')

Constructors

Judgement

Fields

antecedents :: Context a

The antecedents (or hypotheses).

succedent :: Formula a

The succedent (or consequent, or conclusion).

(|-) :: (Ord a, Foldable f) => f (Formula a) -> Formula a -> Judgement a infix 5

Infix judgement constructor. (c |- a) represents the judgement \(c \vdash a\).

>>> [Var 'a', Imp (Var 'a') (Var 'b')] |- Var 'b'
[Var 'a',Imp (Var 'a') (Var 'b')] |- Var 'b'

Note: If c is already a Set, then it is recommended to use (Judgement c a) in favor of (c |- a), since the latter will create a new set and fill it with the values in c.

Pretty-printing

prettyJudgement :: PrettyPrintable a => Judgement a -> String

Get a pretty-printed representation of a judgement.

Examples

Expand

>>> prettyJudgement $ [Var "a", Imp (Var "a") (Var "b")] |- Var "b"
"a, a → b ⊢ b"

Operations on judgements

weakenJudgement :: Ord a => Judgement a -> Formula a -> Judgement a

Weaken a judgement by inserting a formula into its hypotheses.

>>> weakenJudgement ([Var 'a', Imp (Var 'a') (Var 'b')] |- Var 'b') (Var 'c')
[Var 'a',Var 'c',Imp (Var 'a') (Var 'b')] |- Var 'b'