AutoProof.Internal.Utils.Parser

Description

Provides a lightweight, general-purpose parser type and some related combinators.

The API is deliberately nearly a subset of the parsec API. One big difference is the behavior of the Alternative instace; our instance always backtracks, so there is no need for a try function.

The Parser type

data Parser a

The Parser type is a simple state monad with state type String and inner monad Maybe.

Running a parser

parse :: Parser a -> String -> Maybe a

Run a parser on a string, discarding any unparsed suffix

Parser constructors

charIf :: (Char -> Bool) -> Parser Char

Parse a single character if it satisfies a predicate

Examples

Expand

>>> runParser (charIf isDigit) "123"
Just ("23",'1')

>>> runParser (charIf isDigit) "abc"
Nothing

char :: Char -> Parser Char

Parse a specific character.

Examples

Expand

>>> runParser (char 'a') "abc"
Just ("bc",'a')

>>> runParser (char 'a') "def"
Nothing

anyChar :: Parser Char

Parse any character.

Examples

Expand

>>> runParser anyChar "abc"
Just ("bc",'a')

>>> runParser anyChar ""
Nothing

oneOf :: [Char] -> Parser Char

Parse one out of a list of characters.

Examples

Expand

>>> runParser (oneOf ['a', 'b']) "bar"
Just ("ar",'b')

>>> runParser (oneOf ['a', 'b']) "foo"
Nothing

string :: String -> Parser String

Parse a specific string.

Examples

Expand

>>> runParser (string "foo") "foobar"
Just ("bar","foo")

>>> runParser (string "foo") "fobar"
Nothing

spaces :: Parser String

Parse zero or more spaces.

Examples

Expand

>>> runParser spaces "   123"
Just ("123","   ")

>>> runParser spaces "123"
Just ("123","")

>>> runParser spaces ""
Just ("","")

Parser combinators

between :: Parser b -> Parser c -> Parser a -> Parser a

(between open close p) runs open, then p, then close, returning the value parsed by p.

sepBy :: Parser a -> Parser b -> Parser [a]

(sepBy p sep) runs zero or more occurrences of p, separated by sep, returning the list of values parsed by p.

sepBy1 :: Parser a -> Parser b -> Parser [a]

(sepBy1 p sep) runs one or more occurrences of p, separated by sep, returning the list of values parsed by p.

chainl1 :: Parser a -> Parser (a -> a -> a) -> Parser a

(chainl1 p op) runs one or more occurrences of p, separated by op, returning the value obtained by a left-associative application of the functions parsed by op to the values parsed by p.

chainr1 :: Parser a -> Parser (a -> a -> a) -> Parser a

(chainr1 p op) runs one or more occurrences of p, separated by op, returning the value obtained by a right-associative application of the functions parsed by op to the values parsed by p.

notFollowedBy :: Parser a -> Parser ()

(notFollowedBy p) succeeds if and only if p fails.

eof :: Parser ()

(eof) (end-of-file) only succeeds at the end of input.