AutoProof.Classical.CNF

Description

Functions and types related to conjunctive normal forms.

Types

type CNF a = Set (Clause a)

Conjunctive normal form: a conjunction of literals.

type Clause a = Set (Literal a)

A disjunction of literals.

type Literal a = (Bool, a)

Either a propositional variable or its negation, depending on whether the first component is True or False, respectively.

Conversion functions

fromFormula :: Ord a => Formula a -> CNF a

Convert a Formula into conjunctive normal form.

Adapted from Figure 2.6 in

Samuel Mimram (2020) PROGRAM = PROOF.

toFormula :: CNF a -> Formula a

Convert a conjunctive normal form representation of a formula into a formula.